+++ common/src/main/kotlin/com/cobblemon/mod/common/api/storage/pc/PCBox.kt
@@ -89,7 +89,7 @@ open class PCBox(val pc: PCStore) : Iterable<Pokemon> {
     }
 
     fun trackPokemon(pokemon: Pokemon) {
-        pokemon.getChangeObservable()
+        pokemon.changeObservable
             .pipe(
                 stopAfter {
                     val coordinates = it.storeCoordinates.get() ?: return@stopAfter true